requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0
ultralytics>=8.1.0
httpx>=0.25.0
python-multipart>=0.0.6
numpy>=1.24.0
